[refactor] put range code into its own script
